requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.83
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy<=1.0.7
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==5.1.1
easyocr==1.7.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
Flask<=2.3.2
flask-babel==3.1.0
fonttools<=4.39.4
frozenlist==1.3.3
fsspec<=2023.5.0
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
gradio==3.11.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
idna==3.4
imageio==2.29.0
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel<=6.23.1
ipython<=8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client<=8.2.0
jupyter_core<=5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
linkify-it-py==2.0.2
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib<=3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
nest-asyncio<=1.5.6
networkx<=3.0
ninja==1.11.1
numpy<=1.24.1
oauthlib==3.2.2
onnx<=1.14.0
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.14
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas<=2.0.1
paramiko==3.2.0
parso==0.8.3
pdf2docx==0.5.6
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
premailer==3.10.0
prompt-toolkit==3.0.38
protobuf<=3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
PyMuPDF==1.20.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.3
PyWavelets<=1.4.1
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==3.0.0
rarfile==4.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image<=0.20.0
scipy<=1.9.1
seaborn==0.12.2
sentry-sdk==1.23.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.22.0
sympy<=1.11.1
tenacity==8.2.2
tensorboard<=2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile<=2023.4.12
toml==0.10.2
toolz==0.12.0
torch<=2.0.1
torchaudio<=2.0.2
torchvision<=0.15.2
tornado<=6.3.2
tqdm<=4.65.0
traitlets<=5.9.0
tritonclient<=2.33.0
typing_extensions<=4.5.0
tzdata<=2023.3
tzlocal<=5.0.1
uc-micro-py<=1.0.2
ultralytics<=8.0.105
urllib3<=1.26.13
uvicorn<=0.22.0
validators<=0.20.0
visualdl<=2.5.2
watchdog<=3.0.0
wcwidth<=0.2.6
websockets<=11.0.3
Werkzeug<=2.3.4
wget<=3.2
x2paddle<=1.4.1
yarl<=1.9.2
zipp<=3.15.0
zope.event<=4.6
zope.interface<=6.0
